// src/app/types/Equipment.ts
export interface Equipment {
  id: string;
  customerID: string;
  customerName: string;
  siteID: string;
  siteName: string;
  equipmentID: string;
  equipmentName: string;
  typeOfScope: string;
  description: string;
  tagID: string;
  make: string;
  model: string;
  serialNumber: string;
  rangeType: string;
  rangeMin: string;
  rangeMax: string;
  rangeMinTemp: string;
  rangeMaxTemp: string;
  rangeMinPercent: string;
  rangeMaxPercent: string;
  certificateNo: string;
  traceability: string;
}